requirements.txt
rdkit-pypi
onnxruntime
tensorflow
keras-tuner